refact: visualize, summary, signal, io/helpers
