+++ lbm_caiman_python/io.py
@@ -209,3 +209,32 @@ def get_files_ext(base_dir, extension, max_depth):
         if len(file.relative_to(base_path).parts) <= max_depth + 1
     ]
     return matching_files
+
+
+def get_pickle_files(data_path: str | Path) -> list:
+    """
+    Get all .pickle files in a directory and its subdirectories.
+    """
+    files = get_files_ext(data_path, '.pickle', 3)
+    if not files:
+        raise ValueError(f"No .pickle files found in {data_path} or its subdirectories.")
+    return files
+
+
+def get_metrics_path(fname: Path) -> Path:
+    """
+    Get the path to the computed metrics file for a given data file.
+    Assumes the metrics file is to be stored in the same directory as the data file,
+    with the same name stem and a '_metrics.npz' suffix.
+
+    Parameters
+    ----------
+    fname : Path
+        The path to the input data file.
+
+    Returns
+    -------
+    metrics_path : Path
+        The path to the computed metrics file.
+    """
+    return fname.with_stem(fname.stem + '_metrics').with_suffix('.npz')
